Clean up the Room class and related code
 - Removed support for particle systems
   ! A (distant) future commit will add support for GPU particle systems
 * Adjusted Variant string output to include double quotes
 * Changed Instance initializers to use a btVector3 instead of separate doubles
 * Changed Resource and Instance identifiers from int to size_t
 - Temporarily removed support for Room instancemaps
   ! A future commit will add support for full file mapping
 * Moved the Background struct definition to bee/render/background.cpp
 * Moved light rendering from the Room class to bee/render/render.cpp
 * Changed the ViewPort class to use a shared_ptr for its Texture
   ! I am eager to continue replacing raw pointers with smart pointers
 * Changed PathFollowers and TimelineIterator handling code to prefer
   references over pointers
 * Refactored a lot of the Room class
 - Removed is_isometric, background_color, is_background_color_enabled,
   particle_systems, and light_map from the Room class
 - Removed get_background_string() and friends from the Room class

+++ bee/python/instance.cpp
@@ -152,8 +152,8 @@ namespace internal {
 			return nullptr;
 		}
 
-		const std::map<int,Instance*>& instances (obj->get_instances());
-		std::map<int,Instance*>::const_iterator inst (instances.find(self->num));
+		const std::map<size_t,Instance*>& instances (obj->get_instances());
+		std::map<size_t,Instance*>::const_iterator inst (instances.find(self->num));
 		if (inst == instances.end()) {
 			PyErr_SetString(PyExc_ValueError, "the provided Instance id is not valid");
 			return nullptr;
