+++ bee/resource/resource.hpp
@@ -16,7 +16,7 @@ namespace bee {
 			virtual ~Resource() {};
 			virtual int reset() =0;
 			virtual void print() const =0;
-			virtual int get_id() const =0;
+			virtual size_t get_id() const =0;
 			virtual std::string get_name() const {return "";}
 			virtual int load() {return 0;}
 			virtual int free() {return 0;}
